2004-08-02 19:00 UTC+0200 Przemyslaw Czerpak (druzus/at/priv.onet.pl)   * xharbour/include/hbdefs.h    ! include <stdint.h> only for GCC - other developers can make the same      for their compilers if supported
  * xharbour/include/regex.h
  * xharbour/source/rtl/regex.c
   + added memory statistic for regular expressions
   ! fixed memory leak in regex. Each time regcomp is used then
     regree() should be called later to free compiled pcre structure.
     Because we store compiled regex as character itmes (Ron I don't
     like it - it should be IMGO GC pointer item without hits "***"
     hack) I hacked the returned pcre structure and increased the size
     of returned string adding the compiled pcre also. It's working but
     not elegant. Please test it. I'm surprised that for such long time
     knowone noticed it but regex where consuming memory each time they
     were used (regcomp)

+++ xharbour/include/regex.h
@@ -213,7 +213,7 @@ typedef struct real_pcre_extra pcre_extra;
 functions if required. Some magic is required for Win32 DLL; it is null on
 other OS. */
 
-PCRE_DL_IMPORT extern void *(*pcre_malloc)(size_t);
+PCRE_DL_IMPORT extern void *(*pcre_malloc)(ULONG);
 PCRE_DL_IMPORT extern void  (*pcre_free)(void *);
 
 #undef PCRE_DL_IMPORT
